refactor: rename directive appLinkAttributes -> appHtmlAttr
- moved to layout
- feat(app-header): navbarBrand img use appHtmlAttr instead of attr.*
- refactor(app-header): drop unused methods
- refactor(sidebar-nav): directive appLinkAttributes -> appHtmlAttr
