+++ projects/coreui/angular/src/lib/header/app-header.module.ts
@@ -1,6 +1,6 @@
 import { CommonModule} from '@angular/common';
 import { NgModule } from '@angular/core';
-import { LayoutModule } from './../shared/layout/layout.module';
+import { LayoutModule } from '../shared/layout';
 
 import { AppHeaderComponent } from './app-header.component';
