+++ projects/coreui/angular/src/lib/shared/layout/layout.directive.ts
@@ -1,4 +1,4 @@
-import {Directive, HostListener, Inject, Input, OnInit, Renderer2} from '@angular/core';
+import {Directive, ElementRef, HostListener, Inject, Input, OnInit, Renderer2} from '@angular/core';
 import {DOCUMENT} from '@angular/common';
 
 import { asideMenuCssClasses, sidebarCssClasses } from '../classes';
@@ -155,5 +155,48 @@ export class AsideToggleDirective implements OnInit {
     $event.preventDefault();
     const cssClass = this.bp ? `aside-menu-${this.bp}-show` : asideMenuCssClasses[0];
     this.classToggler.toggleClasses(cssClass, asideMenuCssClasses);
+  }
+}
+
+@Directive({
+  selector: '[appHtmlAttr]'
+})
+export class HtmlAttributesDirective implements OnInit {
+  @Input() appHtmlAttr: {[key: string]: string };
+
+  constructor(
+    @Inject(DOCUMENT) private document: any,
+    private renderer: Renderer2,
+    private el: ElementRef
+  ) {}
+
+  ngOnInit() {
+    const attribs = this.appHtmlAttr;
+    for (const attr in attribs) {
+      if (attr === 'style' && typeof(attribs[attr]) === 'object' ) {
+        this.setStyle(attribs[attr]);
+      } else if (attr === 'class') {
+        this.addClass(attribs[attr]);
+      } else {
+        this.setAttrib(attr, attribs[attr]);
+      }
+    }
+  }
+
+  private setStyle(styles) {
+    for (const style in styles) {
+      this.renderer.setStyle(this.el.nativeElement, style, styles[style] );
+    }
+  }
+
+  private addClass(classes) {
+    const classArray = (Array.isArray(classes) ? classes : classes.split(' '));
+    classArray.filter((element) => element.length > 0).forEach(element => {
+      this.renderer.addClass(this.el.nativeElement, element );
+    });
+  }
+
+  private setAttrib(key, value) {
+    this.renderer.setAttribute(this.el.nativeElement, key, value );
   }
 }
